Added a start loading screen
