+++ frontend/CampusSafe/app/_layout.tsx
@@ -6,6 +6,7 @@ export default function RootLayout() {
             <Stack.Screen name="index" options={{ headerShown: false }} />
             <Stack.Screen name="home" options={{ headerShown: false }} />
             <Stack.Screen name="register" options={{ headerShown: false }} />
+            <Stack.Screen name="login" options={{ headerShown: false }} />
             <Stack.Screen name="create_report" options={{ headerShown: false }} />
         </Stack>
     );
