+++ frontend/CampusSafe/components/loading_spinner.tsx
@@ -1,0 +1,8 @@
+import { Colors } from "@/constants/colors";
+import { ActivityIndicator } from "react-native";
+
+const LoadingSpinner = () => {
+    return <ActivityIndicator size="large" color={Colors.primary} />;
+};
+
+export default LoadingSpinner;
