+++ frontend/CampusSafe/constants/colors.ts
@@ -1,0 +1,3 @@
+export const Colors = {
+    primary: "#990000",
+};
